Add base table model
