+++ ui/configure_users.py
@@ -1,7 +1,5 @@
 import copy
-from typing import Dict, List
 
-from PySide2.QtCore import QAbstractTableModel, QModelIndex, Qt
 from PySide2.QtWidgets import (
     QDialog,
     QDialogButtonBox,
@@ -14,62 +12,14 @@ from PySide2.QtWidgets import (
     QVBoxLayout,
 )
 
+from ui.base_table import BaseTableModel
 
-class TableModel(QAbstractTableModel):
-    def __init__(self, data: List[Dict[str, str]]):
-        super().__init__()
-        self._headings = ["name", "email", "facility_user_id", "affiliation"]
-        self._data: List[Dict[str, str]] = copy.deepcopy(data)
 
-    def data(self, index, role=None):
-        if role == Qt.DisplayRole or role == Qt.EditRole:
-            return self._data[index.row()][self._headings[index.column()]]
-
-    def setData(self, index, value, role=None):
-        if role != Qt.EditRole:
-            return False
-
-        row, column_key = index.row(), self._headings[index.column()]
-        self._data[row][column_key] = value
-        return True
-
-    def columnCount(self, parent=None, *args, **kwargs):
-        return len(self._headings)
-
-    def rowCount(self, parent=None, *args, **kwargs):
-        return len(self._data)
-
-    def flags(self, index):
-        return Qt.ItemIsSelectable | Qt.ItemIsEnabled | Qt.ItemIsEditable
-
-    def headerData(self, section, orientation, role=None):
-        if role == Qt.DisplayRole and orientation == Qt.Horizontal:
-            return self._headings[section]
-
-    def add_new_user(self, position):
-        self._data.insert(position, {k: "" for k in self._headings})
-
-    def insertRow(self, position, index=QModelIndex()):
-        self.beginInsertRows(index, position, position)
-        self.add_new_user(position)
-        self.endInsertRows()
-        return True
-
-    def removeRows(self, rows, index=QModelIndex()):
-        for row in sorted(rows, reverse=True):
-            self.beginRemoveRows(QModelIndex(), row, row)
-            del self._data[row]
-            self.endRemoveRows()
-        return True
-
+class UserTableModel(BaseTableModel):
     @property
     def users(self):
         # Ignore users with no name set.
         return [copy.copy(user) for user in self._data if user["name"].strip() != ""]
-
-    @property
-    def num_rows(self):
-        return len(self._data)
 
     def are_users_unique(self):
         users = set()
@@ -87,7 +37,8 @@ class ConfigureUsersDialog(QDialog):
     def __init__(self, users):
         super().__init__()
         self.setWindowTitle("Configure Users")
-        self.model = TableModel(users)
+        headings = ["name", "email", "facility_user_id", "affiliation"]
+        self.model = UserTableModel(headings, users)
 
         self.layout = QVBoxLayout()
         self.table_layout = QHBoxLayout()
