+++ ui/sample_table.py
@@ -1,7 +1,3 @@
-import copy
-from typing import Dict, List
-
-from PySide2.QtCore import QAbstractTableModel, QModelIndex, Qt
 from PySide2.QtWidgets import (
     QDialog,
     QDialogButtonBox,
@@ -13,64 +9,20 @@ from PySide2.QtWidgets import (
     QVBoxLayout,
 )
 
+from ui.base_table import BaseTableModel
 
-class SampleTableModel(QAbstractTableModel):
-    def __init__(self, data: List[Dict[str, str]]):
-        super().__init__()
-        self._headings = ["name", "formula", "number of", "mass/volume", "density"]
-        self._data: List[Dict[str, str]] = copy.deepcopy(data)
 
-    def data(self, index, role=None):
-        if role == Qt.DisplayRole or role == Qt.EditRole:
-            return self._data[index.row()][self._headings[index.column()]]
-
-    def setData(self, index, value, role=None):
-        if role != Qt.EditRole:
-            return False
-
-        row, column_key = index.row(), self._headings[index.column()]
-        self._data[row][column_key] = value
-        return True
-
-    def columnCount(self, parent=None, *args, **kwargs):
-        return len(self._headings)
-
-    def rowCount(self, parent=None, *args, **kwargs):
-        return len(self._data)
-
-    def flags(self, index):
-        return Qt.ItemIsSelectable | Qt.ItemIsEnabled | Qt.ItemIsEditable
-
-    def headerData(self, section, orientation, role=None):
-        if role == Qt.DisplayRole and orientation == Qt.Horizontal:
-            return self._headings[section]
-
-    def add_new_row(self, position):
-        self._data.insert(position, {k: "" for k in self._headings})
-
-    def insertRow(self, position, index=QModelIndex()):
-        self.beginInsertRows(index, position, position)
-        self.add_new_row(position)
-        self.endInsertRows()
-        return True
-
-    def removeRows(self, rows, index=QModelIndex()):
-        for row in sorted(rows, reverse=True):
-            self.beginRemoveRows(QModelIndex(), row, row)
-            del self._data[row]
-            self.endRemoveRows()
-        return True
-
-    @property
-    def num_rows(self):
-        return len(self._data)
+class SampleTableModel(BaseTableModel):
+    def are_samples_unique(self):
+        pass
 
 
 class ConfigureSamplesDialog(QDialog):
     def __init__(self, samples):
         super().__init__()
         self.setWindowTitle("Configure Users")
-        self.model = SampleTableModel(samples)
+        headings = ["name", "formula", "number of", "mass/volume", "density"]
+        self.model = SampleTableModel(headings, samples)
 
         self.layout = QVBoxLayout()
         self.table_layout = QHBoxLayout()
@@ -115,10 +67,10 @@ class ConfigureSamplesDialog(QDialog):
     def _complete_table(self):
         self.button_box.button(self.button_box.Ok).setFocus()
 
-    def _add_user_clicked(self):
+    def _add_sample_clicked(self):
         self.sample_table.model().insertRow(self.model.num_rows)
 
-    def _delete_user_clicked(self):
+    def _delete_sample_clicked(self):
         rows_to_remove = set()
         for index in self.sample_table.selectedIndexes():
             rows_to_remove.add(index.row())
